Fractionnated the codes for easier update
