+++ RedPitayaSDK/inc/configuration.h
@@ -1,0 +1,9 @@
+#ifndef CONFIGURATION_H
+#define CONFIGURATION_H
+
+#define ON 1
+#define OFF 0
+#define TRUE 1
+#define FALSE 0
+
+#endif
