+++ RedPitayaSDK/inc/implementer.h
@@ -1,0 +1,11 @@
+#ifndef IMPLEMENTER_H
+#define IMPLEMENTER_H
+
+#include "common.h"
+#include "control.h"
+#include "udp.h"
+#include "processing.h"
+
+int stop;
+
+#endif
