+++ RedPitayaSDK/inc/processing.h
@@ -1,0 +1,8 @@
+#ifndef PROCESSING_H
+#define PROCESSING_H
+
+#include "common.h"
+
+float* calcul_pixel(char* buffer);
+
+#endif
