+++ RedPitayaSDK/srclib/common.c
@@ -1,0 +1,1 @@
+#include "../inc/common.h"
